Moved write logic from "plugins" to `compiler.js`
