+++ lib/compiler.js
@@ -1,3 +1,4 @@
+const fs              = require('fs-extra')
 const path     = require('upath')
 const async    = require('async')
 const inflect  = require('inflection').inflect
@@ -23,7 +24,24 @@ compiler.start = (sourceDir, targetDir) => {
         context.files,
 
         function(file, done) {
-          return file.write(context, done)
+
+          // TODO: Change wording "isSqueezable" to "isRenderable" to avoid...
+          // ... confusion since ALL files go through `squeeze()`
+          if (file.isSqueezable) {
+            fs.mkdirsSync(path.dirname(file.path.target.full))
+            return file.render(context, (err, output) => {
+              if (err) throw err
+              fs.writeFileSync(file.path.target.full, output)
+              return done(null)
+            })
+          }
+
+          if (!file.isRemovable) {
+            fs.mkdirsSync(path.dirname(file.path.target.full))
+            fs.copySync(file.path.full, file.path.target.full)
+          }
+
+          return done(null)
         },
 
         function(err){
