+++ lib/files/datafile.js
@@ -8,7 +8,8 @@ const File            = require('./file')
 
 module.exports = class datafile extends File {
   constructor(filepath, sourceDir, targetDir) {
-    super(filepath, sourceDir, targetDir)
+    super(filepath, sourceDir)
+    this.setTargetPathsAndHref(sourceDir, targetDir)
   }
 
   // calls back with a boolean indicating whether this class should process the given file.
@@ -16,6 +17,21 @@ module.exports = class datafile extends File {
     const extension = path.extname(filename).toLowerCase()
     const allowedExtensions = ['.json', '.yml', '.yaml']
     return callback(null, allowedExtensions.indexOf(extension) > -1)
+  }
+
+  setTargetPathsAndHref(sourceDir, targetDir) {
+    var targetExt = this.path.ext.toLowerCase()
+    var targetFull = this.path.full
+      .replace(sourceDir, targetDir)
+      .replace(this.path.ext, targetExt)
+
+    this.path.target = {
+      full: targetFull,
+      relative: targetFull.replace(targetDir, '')
+    }
+
+    Object.assign(this.path.target, path.parse(this.path.target.relative))
+    this.href = path.join(process.env.JUS_BASEDIR, this.path.target.relative)
   }
 
   squeeze() {
