+++ lib/files/image.js
@@ -9,7 +9,8 @@ const File            = require('./file')
 
 module.exports = class Image extends File {
   constructor(filepath, sourceDir, targetDir) {
-    super(filepath, sourceDir, targetDir)
+    super(filepath, sourceDir)
+    this.setTargetPathsAndHref(sourceDir, targetDir)
   }
 
   // calls back with a boolean indicating whether this class should process the given file.
@@ -23,6 +24,21 @@ module.exports = class Image extends File {
     return callback(null,
       allowedExtensions.indexOf(extension) > -1 && !isSvgFont
     )
+  }
+
+  setTargetPathsAndHref(sourceDir, targetDir) {
+    var targetExt = this.path.ext.toLowerCase()
+    var targetFull = this.path.full
+      .replace(sourceDir, targetDir)
+      .replace(this.path.ext, targetExt)
+
+    this.path.target = {
+      full: targetFull,
+      relative: targetFull.replace(targetDir, '')
+    }
+
+    Object.assign(this.path.target, path.parse(this.path.target.relative))
+    this.href = path.join(process.env.JUS_BASEDIR, this.path.target.relative)
   }
 
   squeeze() {
