+++ lib/files/layout.js
@@ -8,7 +8,7 @@ const File            = require('./file')
 module.exports = class Layout extends File {
 
   constructor(filepath, sourceDir, targetDir) {
-    super(filepath, sourceDir, targetDir)
+    super(filepath, sourceDir)
   }
 
   // calls back with a boolean indicating whether this class should process the given file.
@@ -20,6 +20,7 @@ module.exports = class Layout extends File {
 
   squeeze() {
     this.squeezed = false
+    this.isRemovable = true // means, will not be written in targetDir
     this.setName()
     this.read()
     this.getFrontmatter()
