+++ lib/files/page.js
@@ -14,7 +14,10 @@ handlebars.registerHelper(lobars)
 
 module.exports = class Page extends File {
   constructor(filepath, sourceDir, targetDir) {
-    super(filepath, sourceDir, targetDir)
+    super(filepath, sourceDir)
+    this.setTargetPathsAndHref(sourceDir, targetDir)
+    // Overwrite `file.href` because "index" is a special routing case
+    this.setIndexHref()
   }
 
   // calls back with a boolean indicating whether this class should process the given file.
@@ -24,8 +27,29 @@ module.exports = class Page extends File {
     return callback(null, allowedExtensions.indexOf(extension) > -1)
   }
 
+  setTargetPathsAndHref(sourceDir, targetDir) {
+    var targetExt = '.html' // force target extension
+    var targetFull = this.path.full
+      .replace(sourceDir, targetDir)
+      .replace(this.path.ext, targetExt)
+
+    this.path.target = {
+      full: targetFull,
+      relative: targetFull.replace(targetDir, '')
+    }
+
+    Object.assign(this.path.target, path.parse(this.path.target.relative))
+    this.href = path.join(process.env.JUS_BASEDIR, this.path.target.relative)
+  }
+
+  setIndexHref() {
+    var tail = this.isIndex ? this.path.dir : path.join(this.path.dir, this.path.name)
+    this.href = path.join(process.env.JUS_BASEDIR, tail)
+  }
+
   squeeze() {
     this.squeezed = false
+    this.isSqueezable = true // means, will be render before written in targetDir
     this.read()
     this.getFrontmatter()
     this.getDOMObject()
@@ -65,11 +89,6 @@ module.exports = class Page extends File {
         delete this[t][image.path.name]
       }
     })
-  }
-
-  setHref() {
-    var tail = this.isIndex ? this.path.dir : path.join(this.path.dir, this.path.name)
-    this.href = path.join(process.env.JUS_BASEDIR, tail)
   }
 
   getFrontmatter() {
