+++ lib/files/stylesheetSTYL.js
@@ -8,7 +8,8 @@ const File            = require('./file')
 module.exports = class StylesheetSTYL extends File {
 
   constructor(filepath, sourceDir, targetDir) {
-    super(filepath, sourceDir, targetDir)
+    super(filepath, sourceDir)
+    this.setTargetPathsAndHref(sourceDir, targetDir)
   }
 
   // calls back with a boolean indicating whether this class should process the given file.
@@ -18,8 +19,24 @@ module.exports = class StylesheetSTYL extends File {
     return callback(null, allowedExtensions.indexOf(extension) > -1)
   }
 
+  setTargetPathsAndHref(sourceDir, targetDir) {
+    var targetExt = '.css' // force target extension
+    var targetFull = this.path.full
+      .replace(sourceDir, targetDir)
+      .replace(this.path.ext, targetExt)
+
+    this.path.target = {
+      full: targetFull,
+      relative: targetFull.replace(targetDir, '')
+    }
+
+    Object.assign(this.path.target, path.parse(this.path.target.relative))
+    this.href = path.join(process.env.JUS_BASEDIR, this.path.target.relative)
+  }
+
   squeeze() {
     this.squeezed = false
+    this.isSqueezable = true // means, will be render before written in targetDir
     this.read()
     this.squeezed = true
   }
