+++ lib/module/basefile.js
@@ -6,7 +6,7 @@ const pluralize       = require('inflection').pluralize
 
 module.exports = class BaseFile  {
 
-  constructor(filepath, sourceDir, targetDir) {
+  constructor(filepath, sourceDir) {
     this.setType()
   }
