+++ lib/module/files/datafile.js
@@ -9,7 +9,7 @@ const File            = require('../basefile')
 
 module.exports = class datafile extends File {
   constructor(filepath, sourceDir, targetDir) {
-    super(filepath, sourceDir, targetDir)
+    super(filepath, sourceDir)
   }
 
   // calls back with a boolean indicating whether this class should process the given file.
