+++ lib/module/files/unknown.js
@@ -5,7 +5,7 @@ const File            = require('../basefile')
 module.exports = class Unknown extends File {
 
   constructor(filepath, sourceDir, targetDir) {
-    super(filepath, sourceDir, targetDir)
+    super(filepath, sourceDir)
   }
 
   // calls back with a boolean indicating whether this class should process the given file.
